pyproject.toml
[tool.poetry]
name = "Baguette-Bot"
version = "0.1.0"
description = ""
authors = ["Yorrick06"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
flask = "^1.1.2"
pymongo = "^3.11.0"
wikipedia = "^1.4.0"
asyncpg = "^0.21.0"
dnspython = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
